GCC 4 fixes.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@25885 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/TermWindow.cpp
@@ -487,7 +487,7 @@ TermWindow::MessageReceived(BMessage *message)
 				
 			font.SetSize(size);	
 			view->SetTermFont(&font);
-			PrefHandler::Default()->setInt32(PREF_HALF_FONT_SIZE, size);
+			PrefHandler::Default()->setInt32(PREF_HALF_FONT_SIZE, (int32)size);
 			
 			_ResizeView(view);
 			break;
